Refactored file structure to reflect basic operation types developed for the refactor (Binary, unary, open/closed. Operation abstractions have also been placed into their own source file OpAbstractions for a convienient single location. I was starting to think that having a specific operation abstraction at the start of a file of Library functions was a bit confusing.
