+++ TG/BatchedOps.go
