+++ TG/Utils.go
@@ -3,11 +3,7 @@ package TG
 // utils.go contains helper functions for this projects
 
 import (
-	"bytes"
 	"math/rand"
-	"os/exec"
-	"strconv"
-	"strings"
 	"time"
 )
 
@@ -93,36 +89,14 @@ func isEqual(a, b []int) bool {
 	return true
 }
 
-func runPythonScript(scriptName string, args ...string) []float64 {
-	// Prepare the command with script name and arguments
-	cmdArgs := append([]string{scriptName}, args...)
-	cmd := exec.Command("python3", cmdArgs...)
+// this helper function checks if two Tensors are vectors of the same dimension
+func SameVectorSize(t1 *Tensor, t2 *Tensor) bool {
 
-	// Buffers to capture the output
-	var out bytes.Buffer
-	var stderr bytes.Buffer
-	cmd.Stdout = &out
-	cmd.Stderr = &stderr
-
-	// Run the command and check for errors
-	err := cmd.Run()
-	if err != nil {
-		panic("Within runPythonScript() --- " + err.Error() + " --- Stderr: " + stderr.String())
+	if len(t1.Shape) != 1 || len(t2.Shape) != 1 { // check if tensors are vectors (1D)
+		return false
 	}
-
-	// Split the output into separate strings
-	output := strings.TrimSpace(out.String())
-	strValues := strings.Split(output, " ")
-
-	// Parse each string as a float64
-	var result []float64
-	for _, str := range strValues {
-		value, err := strconv.ParseFloat(str, 64)
-		if err != nil {
-			panic("Within runPythonScript() --- " + err.Error())
-		}
-		result = append(result, value)
+	if len(t1.Data) != len(t2.Data) { // check if vectors are of same length
+		return false
 	}
-
-	return result
+	return true
 }
