--- TG/op_abstractions.go
@@ -1,5 +1,0 @@
-package TG
-
-/*
-* @notice op_abstractions contains functions and interfaces for generalizing common patterns in tensor operations.
- */
